refactor: split admin.ts into sub-routers, tighten auth rate limits
Task 1: Split 916-line admin.ts into admin/stats.ts, admin/users.ts,
admin/content.ts. admin.ts is now a thin router that mounts them.

Task 2: messages.ts and notifications.ts already had take/limit
pagination — no change needed.

Task 3: Updated verifyOtpRateLimiter to 5 req/15 min (was 10/min);
otpRateLimiter updated to 10 req/15 min. Both skip metromap smoke tests.
tsc --noEmit: 0 errors.
